Button to save new ground-truth image.
The masking technique means the output image is based off the original
ground-truth image. After an aesthetically pleasing change is made, such as
growing the length of the hair, a new ground-truth image must be saved and
the latent space recalculated before a different change is made such as
changing the hair color.